:recycle: Move amount logic to its own component
