+++ frontend/src/pages/ProjectDetails/PaymentActions/PaymentForm/WorkEstimation/index.tsx
@@ -1,26 +1,39 @@
+import { useEffect, useState } from "react";
 import Card from "src/components/Card";
 import { formatMoneyAmount } from "src/utils/money";
 import { useT } from "talkr";
 
 export const BASE_RATE_USD = 500;
+const DEFAULT_NUMBER_OF_DAYS = 2;
 
 interface Props {
-  numberOfDays: number;
-  decreaseNumberOfDays: () => void;
-  increaseNumberOfDays: () => void;
+  onChange: (value: number) => void;
   budget: { initialAmount: number; remainingAmount: number };
   submitDisabled: boolean;
 }
 
-export default function WorkEstimation({
-  numberOfDays,
-  decreaseNumberOfDays,
-  increaseNumberOfDays,
-  budget,
-  submitDisabled,
-}: Props) {
-  const amountToPay = numberOfDays * BASE_RATE_USD;
+export default function WorkEstimation({ onChange, budget, submitDisabled }: Props) {
   const { T } = useT();
+  const [numberOfDays, setNumberOfDays] = useState(DEFAULT_NUMBER_OF_DAYS);
+  const amountToPay = numberOfDays * BASE_RATE_USD;
+
+  useEffect(() => {
+    onChange(amountToPay);
+  }, [amountToPay]);
+
+  const tryIncreaseNumberOfDays = () => {
+    const increment = numberOfDays < 1 ? 0.5 : 1;
+    if (numberOfDays < 20 && budget.remainingAmount - (numberOfDays + increment) * BASE_RATE_USD >= 0) {
+      setNumberOfDays(numberOfDays + increment);
+    }
+  };
+
+  const tryDecreaseNumberOfDays = () => {
+    if (numberOfDays > 0.5) {
+      const decrement = numberOfDays == 1 ? 0.5 : 1;
+      setNumberOfDays(numberOfDays - decrement);
+    }
+  };
   return (
     <Card>
       <div className="flex flex-col gap-10 items-stretch justify-items-center w-full">
@@ -29,10 +42,10 @@ export default function WorkEstimation({
             <span className="font-walsheim font-black">{numberOfDays}</span> <span>{T("payment.form.days")}</span>
           </div>
           <div className="flex flex-row gap-3 text-white items-center">
-            <div className="border rounded-xl w-fit py-2 px-4 hover:cursor-pointer" onClick={decreaseNumberOfDays}>
+            <div className="border rounded-xl w-fit py-2 px-4 hover:cursor-pointer" onClick={tryDecreaseNumberOfDays}>
               {T("payment.form.decrease")}
             </div>
-            <div className="border rounded-xl w-fit py-2 px-4 hover:cursor-pointer" onClick={increaseNumberOfDays}>
+            <div className="border rounded-xl w-fit py-2 px-4 hover:cursor-pointer" onClick={tryIncreaseNumberOfDays}>
               {T("payment.form.increase")}
             </div>
           </div>
